Merged PR 5307: Working Stream-based API implementation for query results
Fix the implementation of Stream for retrieving data from queries.

The implementation requires use of Arc<Mutex<..>>> and Arc<AtomicBoolean<..>> to mutate internal state and track at runtime when borrows are happening. This is needed because the root TokenParser and other state information need to be shared to objects created by the stream, and the caller using stream items will indirectly use the TokenParser to iterate over rows.

The Drop trait is used to notify that shared resources are no longer being borrowed so that poll_next() can borrow resources to check state.

This PR tests the following scenarios:
- A batch of queries with a mix of SELECTs and UPDATEs.
- A batch of queries where not all results are iterated over. Check that the TdsConnection can safely be re-used.
- A batch of queries where the result sets may not get fully consumed. Check that the next result in the batch can be received. Also check the case where the result set that is partially consumed is at the end of the batch, and that the TdsConnection can be reused in this case.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request implements a stream-based API for query results, enhancing the efficiency and flexibility of data retrieval.
- `src/query/result.rs`: Added `DeferredSignal` struct for managing asynchronous signals and implemented new methods for `QueryResultType` to handle expected results.
- Introduced `RowStream` and `QueryResultTypeStream` structs to facilitate streaming of query results.
- Added multiple async test functions to validate the new streaming functionality and ensure proper handling of various query scenarios.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #33027