Merged PR 5415: C++ client using cxx crate
C++ client using cxx crate

Notes:
- You cannot compile a C++ executable with cxx and cc crates. Only C++ libraries.
To test, you need to build and link a separate main.cc.
- When linking with a C++ linker, need to link libcxxbridge1.a, the library generated by build.rs
with the cc crate - libtds.a, and the actual cpp-driver-cxx static library created by Cargo - libcpp_driver_cxx.a
- Types that are opaque to C++ but visible in Rust are in extern "Rust" blocks. These types can be anything,
but they must be defined in the crate containing the bridge (eg cpp-driver-cxx).
- Types that are opaque to Rust but visible in C++ are in extern "C++" blocks.
- Types that are fully exposed to both are in the bridge's module (mod ffi). These must be primitives,
enums with repr(C), or composed of built-in cxx types. They must be defined in the crate with the bridge.

For demo purposes this doesn't provide an API to set fields on ClientContext, blocks on async functions, and just prints query results to the screen.

Usage:
Windows:
1. Run x64 Native Tools Command Prompt (under the Visual Studio 2022 Folder in the start menu)
2. Go to RustProtoType\tds-x\cpp-driver-cxx\src
3. cargo build
4. cl main.cc /I "../../../target/debug/build/cpp-driver-cxx-868819eeff1bfe38/out/cxxbridge/include/cpp-driver-cxx" /I "../../../target/debug/build/cpp-driver-cxx-868819eeff1bfe38/out/cxxbridge/include" ../../../target/debug/cpp_driver_cxx.lib secur32.lib ncrypt.lib advapi32.lib ws2_32.lib userenv.lib crypt32.lib ntdll.lib /MD
5. Run main.exe

Mac:
1. Go to tds-x/cpp-driver-cxx/src directory.
2. cargo build to both generate the C++ libraries and compile the Rust library
3. Link with clang on Mac:
- clang++ main.cc -I. -I ../../../target/cxxbridge/ -std=c++11 -mmacosx-version-min=15.2  -force_load ../../../target/debug/build/cpp-driver-cxx-c5001fa80ea682a8/out/libtds.a  -force_load ../../../target/debug/build/cxx-d23b6ad32e2c801e/out/libcxxbridge1.a ../../../target/debug/libcpp_driver_cxx.a -fexceptions -framework Security -framework CoreFoundation -o test.out
4. Run ./test.out

Linux (using dev/Dockerfile):
1. Run the dockerfile but bind the RustPrototype directory to /src
2. In the container, go to /src/tds-x/cpp-driver-cxx/src
3. export CC=g++
4. cargo build
Verify that the CC variable was reported to be non-empty in the output
5. g++ -I ../../../target/cxxbridge/ main.cc ../../../target/debug/libcpp_driver_cxx.a -lssl -lcrypto -o test.out
6. Run ./test.out

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request introduces a C++ client using the `cxx` crate to enable interoperability between C++ and Rust for TDSCore. It addresses the linked work item by creating a C++ driver shim on top of the Rust core.
- `src/lib.rs`: Defines the Rust side of the C++ client, including connection creation and SQL execution.
- `src/tds_connection.cc` and `src/tds_connection.h`: Implements the C++ side of the client, providing methods to cr...